+++ src/test/java/guitests/DeleteCommandTest.java
@@ -1,8 +1,9 @@
 package guitests;
 
 import org.junit.Test;
-import seedu.address.testutil.TestPerson;
-import seedu.address.testutil.TestUtil;
+
+import seedu.taskell.testutil.TestPerson;
+import seedu.taskell.testutil.TestUtil;
 
 import static org.junit.Assert.assertTrue;
 import static seedu.taskell.logic.commands.DeleteCommand.MESSAGE_DELETE_PERSON_SUCCESS;
