+++ src/test/java/guitests/TaskManagerGuiTest.java
@@ -8,12 +8,13 @@ import org.junit.BeforeClass;
 import org.junit.Rule;
 import org.junit.rules.TestName;
 import org.testfx.api.FxToolkit;
-import seedu.address.TestApp;
-import seedu.address.testutil.TestUtil;
-import seedu.address.testutil.TypicalTestPersons;
+
+import seedu.taskell.TestApp;
 import seedu.taskell.commons.core.EventsCenter;
 import seedu.taskell.model.TaskManager;
 import seedu.taskell.model.person.ReadOnlyPerson;
+import seedu.taskell.testutil.TestUtil;
+import seedu.taskell.testutil.TypicalTestPersons;
 
 import java.util.concurrent.TimeoutException;
