+++ src/test/java/guitests/guihandles/GuiHandle.java
@@ -7,7 +7,7 @@ import javafx.scene.control.TextField;
 import javafx.scene.input.KeyCode;
 import javafx.stage.Stage;
 import javafx.stage.Window;
-import seedu.address.TestApp;
+import seedu.taskell.TestApp;
 import seedu.taskell.commons.core.LogsCenter;
 
 import java.util.logging.Logger;
