+++ src/test/java/guitests/guihandles/MainGuiHandle.java
@@ -2,7 +2,7 @@ package guitests.guihandles;
 
 import guitests.GuiRobot;
 import javafx.stage.Stage;
-import seedu.address.TestApp;
+import seedu.taskell.TestApp;
 
 /**
  * Provides a handle for the main GUI.
