+++ src/test/java/guitests/guihandles/ResultDisplayHandle.java
@@ -3,7 +3,7 @@ package guitests.guihandles;
 import guitests.GuiRobot;
 import javafx.scene.control.TextArea;
 import javafx.stage.Stage;
-import seedu.address.TestApp;
+import seedu.taskell.TestApp;
 
 /**
  * A handler for the ResultDisplay of the UI
