+++ src/test/java/seedu/taskell/commons/core/VersionTest.java
@@ -1,4 +1,4 @@
-package seedu.address.commons.core;
+package seedu.taskell.commons.core;
 
 import org.junit.Rule;
 import org.junit.Test;
