+++ src/test/java/seedu/taskell/commons/util/AppUtilTest.java
@@ -1,4 +1,4 @@
-package seedu.address.commons.util;
+package seedu.taskell.commons.util;
 
 import org.junit.Rule;
 import org.junit.Test;
