+++ src/test/java/seedu/taskell/commons/util/FileUtilTest.java
@@ -1,13 +1,13 @@
-package seedu.address.commons.util;
+package seedu.taskell.commons.util;
 
 
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
 
-import seedu.address.testutil.SerializableTestClass;
-import seedu.address.testutil.TestUtil;
 import seedu.taskell.commons.util.FileUtil;
+import seedu.taskell.testutil.SerializableTestClass;
+import seedu.taskell.testutil.TestUtil;
 
 import java.io.File;
 import java.io.IOException;
