+++ src/test/java/seedu/taskell/commons/util/JsonUtilTest.java
@@ -1,4 +1,4 @@
-package seedu.address.commons.util;
+package seedu.taskell.commons.util;
 
 /**
  * Tests JSON Read and Write
