+++ src/test/java/seedu/taskell/commons/util/XmlUtilTest.java
@@ -1,15 +1,15 @@
-package seedu.address.commons.util;
+package seedu.taskell.commons.util;
 
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
 
-import seedu.address.testutil.TaskManagerBuilder;
-import seedu.address.testutil.TestUtil;
 import seedu.taskell.commons.util.FileUtil;
 import seedu.taskell.commons.util.XmlUtil;
 import seedu.taskell.model.TaskManager;
 import seedu.taskell.storage.XmlSerializableTaskManager;
+import seedu.taskell.testutil.TaskManagerBuilder;
+import seedu.taskell.testutil.TestUtil;
 
 import javax.xml.bind.JAXBException;
 import java.io.File;
