+++ src/test/java/seedu/taskell/storage/JsonUserPrefsStorageTest.java
@@ -1,4 +1,4 @@
-package seedu.address.storage;
+package seedu.taskell.storage;
 
 
 import org.junit.Rule;
