+++ src/test/java/seedu/taskell/testutil/EventsCollector.java
@@ -1,4 +1,4 @@
-package seedu.address.testutil;
+package seedu.taskell.testutil;
 
 import com.google.common.eventbus.Subscribe;
