+++ src/test/java/seedu/taskell/testutil/PersonBuilder.java
@@ -1,4 +1,4 @@
-package seedu.address.testutil;
+package seedu.taskell.testutil;
 
 import seedu.taskell.commons.exceptions.IllegalValueException;
 import seedu.taskell.model.person.*;
