+++ src/test/java/seedu/taskell/testutil/SerializableTestClass.java
@@ -1,4 +1,4 @@
-package seedu.address.testutil;
+package seedu.taskell.testutil;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
