+++ src/test/java/seedu/taskell/testutil/TestPerson.java
@@ -1,4 +1,4 @@
-package seedu.address.testutil;
+package seedu.taskell.testutil;
 
 import seedu.taskell.model.person.*;
 import seedu.taskell.model.tag.UniqueTagList;
